+++ src/app/clients/clients.service.ts
@@ -36,13 +36,18 @@ export class ClientsService {
     return this.http.get('/clients', { params: httpParams });
   }
 
-  getClients(orderBy: string, sortOrder: string, offset: number, limit: number): Observable<any> {
-    const httpParams = new HttpParams()
+  getClients(orderBy: string, sortOrder: string, offset: number, limit: number, subStatus?: string): Observable<any> {
+    let httpParams = new HttpParams()
       .set('offset', offset.toString())
       .set('limit', limit.toString())
       .set('sortOrder', sortOrder)
       .set('orderBy', orderBy)
       .set('includeOfficeHierarchyPath', true.toString());
+
+      if(subStatus) {
+        httpParams = httpParams.set('subStatus', subStatus);
+      }
+
     return this.http.get('/clients', { params: httpParams });
   }
 
@@ -51,15 +56,20 @@ export class ClientsService {
     sortOrder: string,
     offset: number,
     limit: number,
-    countryId: string
+    countryId: string,
+    subStatus?: string,
   ): Observable<any> {
-    const httpParams = new HttpParams()
+    let httpParams = new HttpParams()
       .set('offset', offset.toString())
       .set('limit', limit.toString())
       .set('sortOrder', sortOrder)
       .set('orderBy', orderBy)
       .set('countryId', countryId)
-      .set('includeOfficeHierarchyPath', true.toString());
+      .set('includeOfficeHierarchyPath', true.toString())
+
+      if(subStatus) {
+        httpParams = httpParams.set('subStatus', subStatus);
+      }
     return this.http.get('/clients', { params: httpParams });
   }