+++ RELEASE_NOTES.md
@@ -1,5 +1,8 @@
 # Release Notes
 
+## Version 1.4.2 - Community 1.0.0
+    * Maker Checker 
+        * [CP-3873] - KYC Maker Checker UI Improvements
 ## Version 1.4.1 - Community 1.0.0
 
     * Clients